Moved the Manager class into a separate Swift file.
